requirements.txt
anytree==2.8.0
apparmor==3.0.3
appdirs==1.4.4
application-utility==1.3.2
APScheduler==3.6.3
argcomplete==1.12.3
asn1crypto==1.4.0
attrs==21.4.0
bcrypt==3.2.0
Beaker==1.11.0
beautifulsoup4==4.10.0
bs4==0.0.1
btrfsutil==5.15.1
CacheControl==0.12.6
cached-property==1.5.2
cachetools==4.2.2
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
chrome-gnome-shell==0.0.0
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
commonmark==0.9.1
connexion==2.9.0
contextlib2==0.6.0.post1
cryptography==36.0.1
cupshelpers==1.0
dictdiffer==0.9.0
distlib==0.3.3
distro==1.6.0
dnspython==2.1.0
docker==5.0.3
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
filelock==3.4.2
Flask==1.1.4
greenlet==1.1.2
html5lib==1.1
idna==3.3
importlib-metadata==4.8.1
inflection==0.5.1
isodate==0.6.1
itsdangerous==1.1.0
Jade-Application-Kit==3.5.6
Jinja2==2.11.3
jsonschema==3.2.0
keyutils==0.6
lensfun==0.3.95
LibAppArmor==3.0.3
libtorrent===1.2.15-build-libtorrent-rasterbar-src-libtorrent-rasterbar-1.2.15-bindings-python
louis==3.20.0
lxml==4.6.5
Mako==1.1.6
mallard-ducktype==1.0.2
Markdown==3.3.6
MarkupSafe==2.0.1
meson==0.60.3
more-itertools==8.10.0
msgpack==1.0.3
netsnmp-python==1.0a1
notion==0.0.28
npyscreen==4.10.5
numpy==1.21.5
openapi-schema-validator==0.1.6
openapi-spec-validator==0.3.1
ordered-set==4.0.2
packaging==20.9
pacman-mirrors==4.23.2
pandas==1.3.5
paramiko==2.7.2
pep517==0.12.0
Pillow==8.4.0
pipenv==2021.11.23
platformdirs==2.4.1
ply==3.11
powerline-shell==0.7.0
powerline-status==2.8.2
progress==1.6
pwquality==1.4.4
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.21
pycups==2.0.1
pycurl==7.44.1
Pygments==2.10.0
PyGObject==3.42.0
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.0
PyQtWebEngine==5.15.5
pyrsistent==0.18.0
PySide2==5.15.2
pysmbc==1.0.23
pyTelegramBotAPI==4.2.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-slugify==5.0.2
python-telegram-bot==13.10
pytoml==0.1.21
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
rados==2.0.0
rbd==2.0.0
reportlab==3.6.3
requests==2.26.0
resolvelib==0.5.5
retrying==1.3.3
rgw==2.0.0
shiboken2==5.15.2
shtab==1.5.2
sip==4.19.25
six==1.16.0
smbus==1.1
soupsieve==2.3.1
TBB==0.2
team==1.0
termcolor==1.1.0
text-unidecode==1.3
texttable==1.6.4
tldr==3.0.0
toml==0.10.2
tomli==1.2.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
udiskie==2.4.0
urllib3==1.26.7
virtualenv==20.13.0
virtualenv-clone==0.5.7
waitress==2.0.0
webencodings==0.5.1
websocket-client==0.59.0
Werkzeug==1.0.1
zipp==3.6.0
